quality evaluation score for snippet generation
I added another column for a quality evaluation score that we had in the paper. I forgot to do this the first time. I understand that other follow up studies may or may not do this quality evaluation, but I feel it is necessary to mention that it can be done. Also, updated the dataset link. 

Thanks again for organizing all of this!