Add tagged corruption results on GEC
The system happens to use round-trip translations and Wikipedia revisions in pre-training as described in https://www.aclweb.org/anthology/N19-1333/ They are extracted from public data using public tools/APIs - the parallel data itself is not published but it could be regenerated by anyone (and using them is not the point of the paper). So I believe it classifies as "restricted", but let me know if you think otherwise.